avg model, compute wer, format rsl into paddlespeech.dataset
