+++ paddlespeech/dataset/s2t/format_data.py
@@ -1,5 +1,4 @@
-#!/usr/bin/env python3
-# Copyright (c) 2021 PaddlePaddle Authors. All Rights Reserved.
+# Copyright (c) 2023 PaddlePaddle Authors. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
